+++ packages/botfuel-dialog/tests/dialog-manager/dialog-manager.test.js
@@ -18,7 +18,7 @@
 
 const DialogManager = require('../../src/dialog-manager');
 const Dialog = require('../../src/dialogs/dialog');
-const Intent = require('../../src/nlus/intent');
+const ClassificationResult = require('../../src/nlus/classification-result');
 const MemoryBrain = require('../../src/brains/memory-brain');
 const TestAdapter = require('../../src/adapters/test-adapter');
 const BotTextMessage = require('../../src/messages/bot-text-message');
@@ -33,7 +33,7 @@ const greetingsDialog = { name: 'greetings', entities: [] };
 const thanksDialog = { name: 'thanks', entities: [] };
 const travelDialog = { name: 'travel', entities: [] };
 const travelCancelDialog = { name: 'travel-cancel', entities: [] };
-const intentResolutionDialog = { name: 'intent-resolution', entities: [] };
+const classificationDisambiguationDialog = { name: 'classification-disambiguation', entities: [] };
 
 describe('DialogManager', () => {
   const brain = new MemoryBrain(TEST_CONFIG);
@@ -54,15 +54,15 @@ describe('DialogManager', () => {
 
   test('should not crash when no intent', async () => {
     const adapter = new TestAdapter({});
-    await dm.executeIntents(adapter, { user: TEST_USER }, [], []);
+    await dm.executeClassificationResults(adapter, { user: TEST_USER }, [], []);
     expect(adapter.log).toEqual([new BotTextMessage('Not understood.').toJson(TEST_USER)]);
   });
 
   test('should keep on the stack a dialog which is waiting', async () => {
-    await dm.executeIntents(
+    await dm.executeClassificationResults(
       null,
       { user: TEST_USER },
-      [new Intent({ name: 'waiting', type: Intent.TYPE_INTENT })],
+      [new ClassificationResult({ name: 'waiting', type: ClassificationResult.TYPE_INTENT })],
       [],
     );
     const dialogs = await dm.brain.getDialogs(TEST_USER);
@@ -70,16 +70,16 @@ describe('DialogManager', () => {
   });
 
   test('should not stack the same dialog twice', async () => {
-    await dm.executeIntents(
+    await dm.executeClassificationResults(
       null,
       { user: TEST_USER },
-      [new Intent({ name: 'waiting', type: Intent.TYPE_INTENT })],
+      [new ClassificationResult({ name: 'waiting', type: ClassificationResult.TYPE_INTENT })],
       [],
     );
-    await dm.executeIntents(
+    await dm.executeClassificationResults(
       null,
       { user: TEST_USER },
-      [new Intent({ name: 'waiting', type: Intent.TYPE_INTENT })],
+      [new ClassificationResult({ name: 'waiting', type: ClassificationResult.TYPE_INTENT })],
       [],
     );
     const dialogs = await dm.brain.getDialogs(TEST_USER);
@@ -88,10 +88,10 @@ describe('DialogManager', () => {
 
   test('should empty the stack (1)', async () => {
     const adapter = new TestAdapter({});
-    await dm.executeIntents(
+    await dm.executeClassificationResults(
       adapter,
       { user: TEST_USER },
-      [new Intent({ name: 'default', type: Intent.TYPE_INTENT })],
+      [new ClassificationResult({ name: 'default', type: ClassificationResult.TYPE_INTENT })],
       [],
     );
     const dialogs = await dm.brain.getDialogs(TEST_USER);
@@ -105,28 +105,28 @@ describe('DialogManager', () => {
     expect(dialogs.stack.length).toBe(0);
   });
 
-  test('should call intent resolution dialog when multiple intents detected', async () => {
+  test('should call classification disambiguation dialog when multiple classification results detected', async () => {
     const adapter = new TestAdapter({});
-    await dm.executeIntents(
+    await dm.executeClassificationResults(
       adapter,
       { user: TEST_USER },
       [
-        new Intent({
+        new ClassificationResult({
           name: 'first-intent',
           resolvePrompt: 'first-intent?',
-          type: Intent.TYPE_INTENT,
+          type: ClassificationResult.TYPE_INTENT,
         }),
-        new Intent({
+        new ClassificationResult({
           name: 'second-intent',
           resolvePrompt: 'second-intent?',
-          type: Intent.TYPE_INTENT,
+          type: ClassificationResult.TYPE_INTENT,
         }),
       ],
       [],
     );
     const dialogs = await dm.brain.getDialogs(TEST_USER);
     expect(dialogs.stack.length).toBe(0);
-    expect(dialogs.previous[0].name).toEqual(intentResolutionDialog.name);
+    expect(dialogs.previous[0].name).toEqual(classificationDisambiguationDialog.name);
   });
 
   describe('DialogManager.applyAction', () => {